+++ src/coordinates.h
@@ -2,11 +2,11 @@
 #define MIDDLE_AGES_COORDINATES_H
 
 typedef struct {
-    int x;
-    int y;
+    long x;
+    long y;
 } Coordinates;
 
 int coordinatesCompare(Coordinates a, Coordinates b);
-int coordinatesDistance(Coordinates a, Coordinates b);
+long coordinatesDistance(Coordinates a, Coordinates b);
 
 #endif
